fix manage workshop

+++ frontend/src/pages/Admin/AdminWorkshop/ManageWorkshop.vue
@@ -658,7 +658,7 @@
             <q-btn
               class="tf-capitalize icon-resize"
               color="negative" rounded unelevated label="Remove All Periodic Services"
-              @click="doDeleteAllWorkshopServicesByType(periodicServicesForms)"
+              @click="doDeleteAllWorkshopServicesByType(periodicServicesForms, 'p')"
             />
             <q-btn
               class="tf-capitalize icon-resize"
@@ -669,7 +669,7 @@
             <q-btn
               class="tf-capitalize icon-resize"
               color="negative" rounded unelevated label="Remove All General Services"
-              @click="doDeleteAllWorkshopServicesByType(generalServicesForms)"
+              @click="doDeleteAllWorkshopServicesByType(generalServicesForms, 'g')"
             />
           </div>
           <div class="q-gutter-x-sm">
@@ -1505,29 +1505,46 @@ export default {
         })
       })
     },
-    doDeleteAllWorkshopServicesByType (item) {
+    doDeleteAllWorkshopServicesByType (item, initial) {
       let _this = this
       let firstItem = item[0]
-      deleteWorkshopServiceByStatusAndID(firstItem.workshopDetailID, firstItem.serviceType, _this.accessToken).then(response => {
-        Swal.fire({
-          icon: 'success',
-          title: 'Success',
-          text: response.data.message
-        }) .then(() => {
-          _this.showError = false
-          _this.editCarServiceLoader = false
-          _this.doGetWorkshopDetailByUserID(false, true)
-        })
-      }) .catch((error) => {
-        console.log(error)
-        Swal.fire({
-          icon: 'error',
-          title: 'Error',
-          text: 'Please contact our admin'
-        }) .then(() => {
-          _this.showError = true
-          _this.editCarServiceLoader = false
-        })
+      Swal.fire({
+        icon: 'warning',
+        title: initial == 'p' ? 'Removing all periodic services, are you sure?' : 'Removing all general services, are you sure?',
+        confirmButtonText: 'Yes',
+        confirmButtonColor: '#21a17b',
+        cancelButtonColor: '#d32f2f',
+        cancelButtonText: 'No',
+        showCancelButton: true,
+        reverseButtons: true,
+        customClass: {
+          confirmButton: 'br-25px-i py-5-i px-20-i',
+          cancelButton: 'br-25px-i py-5-i px-20-i'
+        }
+      }) .then((result) => {
+        if(result.isConfirmed){
+          deleteWorkshopServiceByStatusAndID(firstItem.workshopDetailID, firstItem.serviceType, _this.accessToken).then(response => {
+            Swal.fire({
+              icon: 'success',
+              title: 'Success',
+              text: response.data.message
+            }) .then(() => {
+              _this.showError = false
+              _this.editCarServiceLoader = false
+              _this.doGetWorkshopDetailByUserID(false, true)
+            })
+          }) .catch((error) => {
+            console.log(error)
+            Swal.fire({
+              icon: 'error',
+              title: 'Error',
+              text: 'Please contact our admin'
+            }) .then(() => {
+              _this.showError = true
+              _this.editCarServiceLoader = false
+            })
+          })
+        }
       })
     },
     addGalleryForm () {
